create a get default pokemon route
